move files, remove unused exports and update readme
