+++ app/actions/patchDocument.ts
@@ -1,16 +1,16 @@
 'use server';
+import patchableEntityMapper from '@/tina/patchable-entity/patchableEntityMapper';
 import type { PatchParcel, Patches } from '@/types/index';
 import { patchDocument } from 'docx';
 import { readFileSync, writeFileSync } from 'fs';
 import path from 'path';
-import entityMapper from 'tina/entityMapper';
 import { IS_RUNNING_LOCALLY } from '../../utils/constants';
 import { staticFilePath } from '../../utils/path';
 import { supabaseUpsert } from '../../utils/supabase';
 
 export default async (patchParcel: PatchParcel) => {
   const { entity, filename, placeholders, skeleton } = patchParcel;
-  const givenEntity = entityMapper(entity?.__typename!);
+  const givenEntity = patchableEntityMapper[entity?.__typename!];
   const patches: Patches = givenEntity.patches(entity, placeholders);
   const file = readFileSync(staticFilePath(skeleton));
