+++ tina/patchable-entity/patchableEntityMapper.ts
@@ -1,0 +1,6 @@
+import type { PatchableEntity } from '@/types/index';
+import audit from './variants/audit';
+
+export default {
+  Audit: audit,
+} as Record<string, PatchableEntity>;
