+++ tina/patchable-entity/variants/audit/handle/patch.ts
