+++ utils/constants/index.ts
@@ -7,16 +7,10 @@ export const LETTERS_NUMBERS_HYPEN_BLANK_REGEX = /^[a-zA-Z0-9-\s]+$/;
 export const IS_RUNNING_LOCALLY =
   !!process && process.env.NODE_ENV === 'development';
 
-export const SUPABASE_CLIENT = process.env.NEXT_PUBLIC_SUPABASE_URL && process.env.NEXT_PUBLIC_SUPABASE_KEY ? createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL,
-  process.env.NEXT_PUBLIC_SUPABASE_KEY
-) : undefined;
-
-export const TINA_URL =
-  !!process &&
-  process.env.NEXT_PUBLIC_TINA_URL &&
-  process.env.NEXT_PUBLIC_TINA_URL.includes('localhost')
-    ? process.env.NEXT_PUBLIC_TINA_URL
-    : `${process.env.NEXT_PUBLIC_TINA_URL}${
-        process.env.HEAD || process.env.VERCEL_GIT_COMMIT_REF || 'main'
-      }`;
+export const SUPABASE_CLIENT =
+  process.env.NEXT_PUBLIC_SUPABASE_URL && process.env.NEXT_PUBLIC_SUPABASE_KEY
+    ? createClient(
+        process.env.NEXT_PUBLIC_SUPABASE_URL,
+        process.env.NEXT_PUBLIC_SUPABASE_KEY
+      )
+    : undefined;
